+++ edflow/iterators/template_iterator.py
@@ -12,6 +12,8 @@ from edflow.util import (
     get_obj_from_str,
     get_str_from_obj,
 )
+import numpy as np
+from edflow.iterators.batches import deep_lod2dol
 
 
 class TemplateIterator(PyHookedModelIterator):
@@ -242,3 +244,32 @@ class TemplateIterator(PyHookedModelIterator):
         written to the log file and stdout. Outputs of 'eval_op' are written
         into the project's eval folder to be evaluated with `edeval`."""
         raise NotImplemented()
+
+    def get_fixed_examples(self, which_set="train"):
+        """collect fixed examples from dataset with names.
+        """
+        if not hasattr(self, "fixed_examples"):
+            fixed_random_indices = np.random.RandomState(1).choice(
+                len(self.dataset), self.config["batch_size"]
+            )
+
+            default_value = {
+                "train": fixed_random_indices,
+                "validation": fixed_random_indices,
+            }
+            fixed_example_indices = set_default(
+                self.config, "fixed_example_indices", default_value
+            )
+
+            fixed_examples_train = [
+                self.datasets["train"][i] for i in fixed_example_indices["train"]
+            ]
+            fixed_examples_val = [
+                self.datasets["validation"][i]
+                for i in fixed_example_indices["validation"]
+            ]
+            self.fixed_examples = {
+                "train": deep_lod2dol(fixed_examples_train),
+                "validation": deep_lod2dol(fixed_examples_val),
+            }
+        return self.fixed_examples[which_set]